Move subscribe/unsubscribe functions to PubnubClient. Move EE related functions to internal utilities.
